requirements.txt
# Core (Discord)
discord.py
# Menus
git+https://github.com/Rapptz/discord-ext-menus

# Other
pendulum

# Database
motor
dnspython

# Error Handling/Debugging
sentry-sdk
psutil
jishaku